Division Types Master
Add New V_Division.py,S_Division.py
Remove code on Parties view and Serializer
