+++ FoodERP/FoodERPApp/Views/V_Parties.py
@@ -45,22 +45,7 @@ class GetPartyTypeByDivisionTypeID(CreateAPIView):
         except Exception as e:
             return JsonResponse({'StatusCode': 400, 'Status': True, 'Message':  Exception(e), 'Data':[]})            
             
-class M_DivisionTypeView(CreateAPIView):
-    
-    permission_classes = (IsAuthenticated,)
-    authentication__Class = JSONWebTokenAuthentication
-
-    @transaction.atomic()
-    def get(self, request):
-        try:
-            with transaction.atomic():
-                M_DivisionType_data = M_DivisionType.objects.all()
-                if M_DivisionType_data.exists():
-                    M_DivisionType_serializer = M_DivisionTypeSerializer(M_DivisionType_data, many=True)
-                    return JsonResponse({'StatusCode': 200, 'Status': True,'Message': '','Data':M_DivisionType_serializer.data })
-                return JsonResponse({'StatusCode': 204, 'Status': True,'Message':  'Division Not available', 'Data': []})    
-        except Exception as e:
-            return JsonResponse({'StatusCode': 400, 'Status': True, 'Message':  Exception(e), 'Data':[]})            
+         
 
 class M_PartiesView(CreateAPIView):
